apps/chronoperates-api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chronoperates-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.66.0",
    "fastapi>=0.116.1",
    "icalendar>=6.3.1",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.15.0",
    "ruff>=0.12.12",
]
